Feat: refactor test
